:truck: Move ResponsivityFallback
