+++ frontend/src/components/Layout/index.tsx
@@ -1,6 +1,6 @@
 import classNames from "classnames";
 import { Outlet, useLocation } from "react-router-dom";
-import ResponsivityFallback from "../ResponsivityFallback";
+import ResponsivityFallback from "./ResponsivityFallback";
 import { Toaster } from "./Toaster";
 
 import Header from "./Header";
